# backend/public/openapi.yaml
openapi: 3.1.0
info:
  title: SaborExpress API
  version: 0.3.0
  summary: API REST para gestão acadêmica de restaurante.
  license:
    name: MIT
    identifier: MIT
  description: >
    Documentação OpenAPI da API atual do SaborExpress. A especificação descreve
    apenas as rotas implementadas no backend Laravel.
servers:
  - url: /api
    description: Servidor da aplicação atual
...
security:
  - bearerAuth: []
paths:
...
  /orders/{order}/request-closing:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags:
        - Orders
      summary: Solicita o fechamento de uma comanda.
      operationId: requestOrderClosing
      description: Exige perfil `administrator`, `waiter` ou `cashier`. Transiciona a comanda de `Aberta` para `Fechamento` e coloca a mesa em `Fechamento`.
      responses:
        '200':
          description: Comanda em fechamento.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /orders/{order}/cancel:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags:
        - Orders
      summary: Cancela uma comanda.
      operationId: cancelOrder
      description: Exige perfil `administrator`, `waiter` ou `cashier`. Comandas pagas ou com pagamento pago nao podem ser canceladas.
      responses:
        '200':
          description: Comanda cancelada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /orders/{order}/payments:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags:
        - Payments
        - Orders
      summary: Registra o pagamento integral de uma comanda.
      operationId: registerOrderPayment
      description: Exige perfil `administrator` ou `cashier`. A comanda precisa estar em `Fechamento` e o valor informado deve bater com o total atual.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPaymentRequest'
      responses:
        '201':
          description: Pagamento registrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /orders/{order}/items:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    post:
      tags:
        - Orders
        - Order Items
      summary: Adiciona um item à comanda.
      operationId: addOrderItem
      description: Exige perfil `administrator`, `waiter` ou `cashier`. A comanda precisa estar `Aberta`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddOrderItemRequest'
      responses:
        '201':
          description: Item adicionado e comanda recalculada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResourceResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum token
  parameters:
...
    OrderId:
      name: order
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    Unauthenticated:
      description: Usuário não autenticado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Forbidden:
      description: Usuário sem permissão para acessar o recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    NotFound:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    Conflict:
      description: Conflito com o estado atual do recurso.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
...
    ValidationError:
      description: Erro de validação.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    User:
      type: object
      required:
        - id
        - name
        - email
        - role
        - is_active
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - administrator
            - waiter
            - kitchen
            - cashier
            - delivery
            - client
        is_active:
          type: boolean
    Category:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
...
    Product:
      type: object
      required:
        - id
        - category_id
        - name
        - price
        - is_available
      properties:
        id:
          type: integer
        category_id:
          type: integer
        name:
          type: string
        description:
          type:
            - string
            - 'null'
        image_url:
          type:
            - string
            - 'null'
          format: uri
        price:
          type: string
          example: '42.90'
        is_available:
          type: boolean
        stock_quantity:
          type:
            - integer
            - 'null'
          minimum: 0
        category:
          $ref: '#/components/schemas/Category'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
...
    Table:
      type: object
      required:
        - id
        - number
        - capacity
        - status
      properties:
        id:
          type: integer
        number:
          type: string
        capacity:
          type: integer
        status:
          $ref: '#/components/schemas/TableStatus'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
...
    TableStatus:
      type: string
      enum:
        - Livre
        - Ocupada
        - Reservada
        - Fechamento
        - Limpeza
...
    Order:
      type: object
      required:
        - id
        - status
        - type
        - total_amount
      properties:
        id:
          type: integer
        table_id:
          type:
            - integer
            - 'null'
        user_id:
          type:
            - integer
            - 'null'
        status:
          $ref: '#/components/schemas/OrderStatus'
        type:
          type: string
          enum:
            - Mesa
            - Delivery
            - Takeout
        customer_name:
          type:
            - string
            - 'null'
        customer_phone:
          type:
            - string
            - 'null'
        delivery_address:
          type:
            - string
            - 'null'
        delivery_status:
          type:
            - string
            - 'null'
          enum:
            - Aguardando
            - Em Rota
            - Entregue
            - null
        total_amount:
          type: string
          example: '84.50'
        discount:
          type: string
          example: '0.00'
        service_fee:
          type: string
          example: '0.00'
        table:
          $ref: '#/components/schemas/Table'
        user:
          $ref: '#/components/schemas/User'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
    OrderStatus:
      type: string
      enum:
        - Aberta
        - Fechamento
        - Paga
        - Cancelada
    OrderItem:
      type: object
      required:
        - id
        - order_id
        - product_id
        - quantity
        - unit_price
        - line_total
        - status
      properties:
        id:
          type: integer
        order_id:
          type: integer
        product_id:
          type: integer
        quantity:
          type: integer
        unit_price:
          type: string
          example: '42.90'
        line_total:
          type: string
          example: '85.80'
        notes:
          type:
            - string
            - 'null'
        status:
          $ref: '#/components/schemas/OrderItemStatus'
        product:
          $ref: '#/components/schemas/Product'
        order:
          $ref: '#/components/schemas/Order'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
    OrderItemStatus:
      type: string
      enum:
        - Pendente
        - Em Preparo
        - Pronto
        - Entregue
        - Cancelado
...
    Payment:
      type: object
      required:
        - id
        - order_id
        - user_id
        - method
        - status
        - amount
      properties:
        id:
          type: integer
        order_id:
          type: integer
        user_id:
          type: integer
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          type: string
          example: '84.50'
        paid_at:
          type:
            - string
            - 'null'
          format: date-time
        notes:
          type:
            - string
            - 'null'
        order:
          $ref: '#/components/schemas/Order'
        user:
          $ref: '#/components/schemas/User'
        created_at:
          type:
            - string
            - 'null'
          format: date-time
        updated_at:
          type:
            - string
            - 'null'
          format: date-time
    PaymentMethod:
      type: string
      enum:
        - Pix
        - Cartao
        - Dinheiro
...
    PaymentStatus:
      type: string
      enum:
        - Pendente
        - Pago
        - Cancelado
        - Estornado
...
    AddOrderItemRequest:
      type: object
      required:
        - product_id
        - quantity
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
          minimum: 1
        notes:
          type:
            - string
            - 'null'
          maxLength: 1000
...
    RegisterPaymentRequest:
      type: object
      required:
        - method
        - amount
      properties:
        method:
          $ref: '#/components/schemas/PaymentMethod'
        amount:
          type: number
          format: float
          minimum: 0
          example: 84.5
        notes:
          type:
            - string
            - 'null'
          maxLength: 1000
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    OrderResourceResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Order'
...
    PaymentResourceResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/Payment'
...